Wrote a function for adding or subtracting 330 minutes from the time. Used that function in places where needed. Solved the console error which was caused when user the edits a blog. Re-organized blog card related files in the folder structure.
