+++ src/Components/drawer/blogs/BlogCard/BlogCard.scss
