+++ src/Components/drawer/blogs/blogsList/index.jsx
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { useQuery } from '@apollo/react-hooks';
+import { useLocation, useHistory } from 'react-router-dom';
 import { GET_ALL_BLOGS } from '../../../../graphql/queries';
 import SomethingWentWrong from '../../../common/SomethingWentWrong/index';
 import useSessionExpired from '../../../../customHooks/useSessionExpired';
@@ -7,10 +8,24 @@ import AllBlogsPage from './AllBlogsPage';
 import PageCountDisplayer from '../../../common/PageCountDisplayer';
 import Spinner from '../../../common/Spinner/index';
 import useActivePageState from '../../../../customHooks/useAcitvePageState';
+import CustomSnackbar from '../../../common/Snackbar/index';
 
 const AllBlogsPageContainer = () => {
   const { redirectOnSessionExpiredBeforeRender, isSessionExpired } = useSessionExpired();
   const limit = 15;
+  const location = useLocation();
+  const { state } = location;
+  const history = useHistory();
+  const [snackbarMessage, setSnackbarMessage] = useState(state && state.snackbarMessage ? state.snackbarMessage : '');
+
+  // This useEffect logic removes the snackbar message from the state
+  // and thereby avoiding the snackbar message being showed everytime the user visits the website
+  useEffect(() => {
+    if (state && state.snackbarMessage) {
+      delete state.snackbarMessage;
+      history.replace({ location, state });
+    }
+  }, [history, location, state]);
   const activePageNumber = useActivePageState();
   const {
     loading, error, data, networkStatus,
@@ -35,6 +50,10 @@ const AllBlogsPageContainer = () => {
             activePageNumber={activePageNumber}
           />
         </div>
+        <CustomSnackbar
+          setSnackbarMessage={setSnackbarMessage}
+          snackbarMessage={snackbarMessage}
+        />
       </div>
     );
   }
